requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-swagger==1.0.16
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
catboost==1.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
graphviz==0.17
idna==3.3
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
lightgbm==3.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
motor==2.5.1
multidict==5.2.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.3.1
pymongo==3.12.1
pyparsing==3.0.3
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.5.3
requests==2.26.0
rq==1.10.0
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.0.0
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarl==1.7.0
